feat: created userController class with all method handling for user related requests
